feat: redefinindo as rotas da aplicação
