+++ src/config/navigator.tsx
@@ -1,20 +1,13 @@
 import React from 'react';
 import {NavigationContainer} from '@react-navigation/native';
 import {createStackNavigator} from '@react-navigation/stack';
-import * as Features from '../features';
 
-const Stack = createStackNavigator();
+import MainRoutes from '../routes/main';
 
 const Navigator = () => {
   return (
     <NavigationContainer>
-      <Stack.Navigator initialRouteName="Home">
-        <Stack.Screen
-          name="Home"
-          component={Features.HomeFeature.Config.Routes}
-          options={{headerShown: false}}
-        />
-      </Stack.Navigator>
+      <MainRoutes />
     </NavigationContainer>
   );
 };
