--- src/routes/index.ts
@@ -1,3 +1,0 @@
-import Tabs from './tabs';
-
-export {Tabs};
